Move Rail to top level
