+++ rail/docker-compose.yaml
@@ -1,10 +1,10 @@
 version: '3'
 services:
-  spire:
+  rail:
     image: ghcr.io/chronicleprotocol/rail:0.0.0-dev.0
     build:
-      context: "../.."
-      dockerfile: "./cmd/Dockerfile"
+      context: "."
+      dockerfile: "./Dockerfile"
       args:
         APP_NAME: "rail"
         APP_VERSION: "0.0.0-dev.0"
